pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "desksearch"
version = "0.1.0"
description = "Private semantic search engine for your local files"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Shuai Wang", email = "shuai.wang@uq.edu.au" },
]
keywords = ["search", "semantic", "local", "privacy", "retrieval"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Text Processing :: Indexing",
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "onnxruntime>=1.17.0",
    "tokenizers>=0.15.0",
    "huggingface-hub>=0.20.0",
    "faiss-cpu>=1.7.4",
    "tantivy>=0.22.0",
    "watchdog>=4.0.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.6.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "aiofiles>=23.2.0",
    "pymupdf>=1.24.0",
]

[project.optional-dependencies]
full = ["sentence-transformers>=3.0.0"]
ocr = ["pytesseract>=0.3.10", "Pillow>=10.0.0"]
desktop = ["pystray>=0.19.0", "Pillow>=10.0.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "httpx>=0.27.0"]

[project.scripts]
desksearch = "desksearch.__main__:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/desksearch"]
